+++ packages/demo-cli/test/suite.js
@@ -2,7 +2,7 @@
 
 const expect = require('expect.js')
 
-describe('porter-demo', function() {
+describe('demo-cli', function() {
   it('should be able to load prismjs', function() {
     expect(typeof require('prismjs').highlightAll).to.eql('function')
   })