+++ packages/demo-typescript/components/test/suite.ts
@@ -3,7 +3,7 @@
 import * as expect from 'expect.js'
 import * as Prism from 'prismjs'
 
-describe('porter-ts-demo', function() {
+describe('demo-typescript', function() {
   it('should be able to import prismjs', function() {
     expect(Prism.highlightAll).to.be.a(Function)
   })